Update NFDRS_percentile_trend_analysis_v5.py
Updated to new WIMS XML API format. Added format handler for staffing level ("sl") field to properly retain only the integer portion during the update. The data schema is incorrect for this field in the service (is integer, should be text). For now, any modifiers (- or +) are dropped. The service will be revised in the future to use the proper data type.